Split dedup/decon code
