+++ lib/marin/src/marin/processing/classification/config/quickstart_dedupe.yaml
@@ -1,8 +1,4 @@
 input_path: gs://marin-us-central2/documents/hello_world_fw/v1.0/quickstart/
 output_path: gs://marin-us-central2/attributes/hello_world_fw/v1.0/quickstart_duplicatesv4/
 attribute_name: duplicate_text
-min_length: 0
-min_words: 0
-estimated_doc_count: 1000000
-false_positive_rate: 0.001
 processes: 1
